Pretty much done backend

+++ p/javaBackend/src/main/java/com/msglearning/javabackend/services/GameService.java
@@ -1,14 +1,14 @@
 package com.msglearning.javabackend.services;
 
-import com.msglearning.javabackend.converters.GameConverter;
 import com.msglearning.javabackend.entity.Game;
+import com.msglearning.javabackend.entity.GameGenre;
 import com.msglearning.javabackend.entity.Purchase;
 import com.msglearning.javabackend.entity.Rating;
+import com.msglearning.javabackend.repositories.GameGenreRepository;
 import com.msglearning.javabackend.repositories.GameRepository;
 import com.msglearning.javabackend.repositories.PurchaseRepository;
 import com.msglearning.javabackend.repositories.RatingRepository;
-import com.msglearning.javabackend.to.GameTO;
-import org.springframework.http.HttpStatus;
+import javassist.NotFoundException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -19,28 +19,28 @@ import java.util.Optional;
 public class GameService {
 
     private final GameRepository gameRepository;
-    private final GameConverter gameConverter;
-    private final PurchaseRepository purchaseRepository;
+
     private final RatingRepository ratingRepository;
+    private final PurchaseRepository purchaseRepository;
+    private final GameGenreRepository gameGenreRepository;
 
-    public GameService(GameRepository gameRepository, GameConverter gameConverter, PurchaseRepository purchaseRepository, RatingRepository ratingRepository) {
+    public GameService(GameRepository gameRepository, RatingRepository ratingRepository,
+                       PurchaseRepository purchaseRepository, GameGenreRepository gameGenreRepository) {
         this.gameRepository = gameRepository;
-        this.gameConverter = gameConverter;
-        this.purchaseRepository = purchaseRepository;
         this.ratingRepository = ratingRepository;
+        this.purchaseRepository = purchaseRepository;
+        this.gameGenreRepository = gameGenreRepository;
     }
 
-    public List<GameTO> getAllGames() {
-        List<Game> games = gameRepository.findAll();
-        return gameConverter.convertToTOList(games);
+    public List<Game> getAllGames() {
+        return gameRepository.findAll();
     }
-
-    public ResponseEntity<GameTO> getGameById(Long id) {
+    
+    public ResponseEntity<Game> getGameById(Long id) {
         Optional<Game> optionalGame = gameRepository.findById(id);
         if (optionalGame.isPresent()) {
             Game game = optionalGame.get();
-            GameTO gameTO = gameConverter.convertToTO(game);
-            return ResponseEntity.ok(gameTO);
+            return ResponseEntity.ok(game);
         } else {
             return ResponseEntity.notFound().build();
         }
@@ -50,45 +50,54 @@ public class GameService {
         return gameRepository.existsByName(name);
     }
 
-    public ResponseEntity<GameTO> saveGame(GameTO gameTO) {
-        Game game = gameConverter.convertToEntity(gameTO);
-        Game savedGame = gameRepository.save(game);
-        GameTO savedGameTO = gameConverter.convertToTO(savedGame);
-        return ResponseEntity.status(HttpStatus.CREATED).body(savedGameTO);
-    }
 
-    public ResponseEntity<GameTO> updateGame(Long id, GameTO gameTO) {
+    //TODO: saveGame should receive a Game object not GameTO
+    public ResponseEntity<Game> saveGame(Game game) {
+        Game newGame = gameRepository.save(game);
+        return ResponseEntity.ok(newGame);
+    }
+    
+    public ResponseEntity<Game> updateGame(Long id, Game newGame) {
         Optional<Game> optionalGame = gameRepository.findById(id);
         if (optionalGame.isPresent()) {
             Game existingGame = optionalGame.get();
             // Update game attributes
-            existingGame.setName(gameTO.getName());
-            existingGame.setDescription(gameTO.getDescription());
-            existingGame.setImageUrl(gameTO.getImageUrl());
-            existingGame.setPublishDate(gameTO.getPublishDate());
+            existingGame.setName(newGame.getName());
+            existingGame.setDescription(newGame.getDescription());
+            existingGame.setImageUrl(newGame.getImageUrl());
+            existingGame.setPublishDate(newGame.getPublishDate());
+            existingGame.setPrice(newGame.getPrice());
+
+
+            //Save the newGame
             Game updatedGame = gameRepository.save(existingGame);
-            GameTO updatedGameTO = gameConverter.convertToTO(updatedGame);
-            return ResponseEntity.ok(updatedGameTO);
+            return ResponseEntity.ok(updatedGame);
         } else {
             return ResponseEntity.notFound().build();
         }
     }
 
-    public void deleteGame(Long id) {
-        List<Purchase> purchases = purchaseRepository.findByGameId(id);
+
+    public void deleteGame(Long id) throws NotFoundException{
+        //Check if game exists in table
+        if(!gameRepository.existsById(id))
+            throw new NotFoundException("Game id does not exist");
+
+        //Delete gameGenre connections which contain the game
+        List<GameGenre> gameGenres = gameGenreRepository.findByGameId(id);
+        gameGenres.forEach(gameGenre -> gameGenreRepository.deleteById(gameGenre.getId()));
+
+        //Delete purchases which contain the game
+            List<Purchase> purchases = purchaseRepository.findByGameId(id);
+            purchases.forEach(purchase -> purchaseRepository.deleteById(purchase.getId()));
+
+        // Delete all ratings related to the game
         List<Rating> ratings = ratingRepository.findByGameId(id);
-
-        // Delete all purchases related to the game
-        for (Purchase purchase : purchases) {
-            purchaseRepository.deleteById(purchase.getId());
-        }
-
-        // Delete all ratings related to the user
         for (Rating rating : ratings) {
             ratingRepository.deleteById(rating.getId());
         }
 
+        //Delete the game
         gameRepository.deleteById(id);
     }
 }
-
